kernel: Allow building dtbo.img
taimen uses the dtbo.img as an DTB Overlay in order to use the same defconfig for both, taimen and wahoo.
The dtbo.img is built with "make dtbo.img". Allow automatically building it when setting TARGET_NEEDS_DTBOIMAGE to true, or by manually calling make dtboimage.

Thanks to luca020400 for help with the kernel.mk logic

Change-Id: Iff63b8e17756f8ee524ed429ec73b66cdc3c3052
Signed-off-by: phenomenall <venkysaihot@gmail.com>